+++ api/src/main/java/org/allaymc/api/world/sound/SimpleSound.java
@@ -78,14 +78,14 @@ public enum SimpleSound implements Sound {
     POTION_BREWED,
     // LECTERN_BOOK_PLACE is a sound played when a book is placed in a lectern.
     LECTERN_BOOK_PLACE,
-    // SIGN_WAXED is a sound played when a sign is waxed.
-    SIGN_WAXED,
+    // WAXED is a sound played when a copper block or sign is waxed.
+    WAXED,
+    // WAX_REMOVED is a sound played when wax is removed from a copper block or sign.
+    WAX_REMOVED,
     // WAXED_SIGN_FAILED_INTERACTION is a sound played when a player tries to interact with a waxed sign.
     WAXED_SIGN_FAILED_INTERACTION,
     // GLOW_INK_SAC_USED is a sound played when a player used glow ink sac in a sign.
     GLOW_INK_SAC_USED,
-    // WAX_REMOVED is a sound played when wax is removed from a block.
-    WAX_REMOVED,
     // COPPER_SCRAPED is a sound played when a player scrapes a copper block to reduce its oxidation level.
     COPPER_SCRAPED,
     // DECORATED_POT_INSERT_FAILED is a sound played when an item fails to be inserted into a decorated pot.